+++ src/Sentry.Unity.Android/SentryNativeAndroid.cs
@@ -3,7 +3,6 @@ using Sentry.Extensibility;
 using Sentry.Unity.Integrations;
 using Sentry.Unity.NativeUtils;
 using UnityEngine;
-using UnityEngine.Analytics;
 
 namespace Sentry.Unity.Android;
 
@@ -113,25 +112,14 @@ public static class SentryNativeAndroid
 
         Logger?.LogDebug("Fetching installation ID");
 
-        options.DefaultUserId = SentryJava.GetInstallationId();
-        if (string.IsNullOrEmpty(options.DefaultUserId))
+        var installationId = SentryJava.GetInstallationId();
+        if (!string.IsNullOrEmpty(installationId))
         {
-            // In case we can't get an installation ID we create one and sync that down to the native layer
-            Logger?.LogDebug(
-                "Failed to fetch 'Installation ID' from the native SDK. Creating new 'Default User ID'.");
-
-            // We fall back to Unity's Analytics Session Info: https://docs.unity3d.com/ScriptReference/Analytics.AnalyticsSessionInfo-userId.html
-            // It's a randomly generated GUID that gets created immediately after installation helping
-            // to identify the same instance of the game
-            options.DefaultUserId = AnalyticsSessionInfo.userId;
-            if (options.DefaultUserId is not null)
-            {
-                options.ScopeObserver.SetUser(new SentryUser { Id = options.DefaultUserId });
-            }
-            else
-            {
-                Logger?.LogDebug("Failed to create new 'Default User ID'.");
-            }
+            options.DefaultUserId = installationId;
+        }
+        else
+        {
+            Logger?.LogDebug("Failed to fetch 'Installation ID' from the native SDK.");
         }
 
         Logger?.LogInfo("Successfully configured the Android SDK");