+++ src/Sentry.Unity.Native/SentryNative.cs
@@ -3,7 +3,6 @@ using Sentry.Extensibility;
 using Sentry.Unity.Integrations;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Analytics;
 
 namespace Sentry.Unity.Native;
 
@@ -67,12 +66,6 @@ public static class SentryNative
         options.NativeContextWriter = new NativeContextWriter();
         options.NativeDebugImageProvider = new NativeDebugImageProvider();
 
-        options.DefaultUserId = GetInstallationId();
-        if (options.DefaultUserId is not null)
-        {
-            options.ScopeObserver.SetUser(new SentryUser { Id = options.DefaultUserId });
-        }
-
         // Note: we must actually call the function now and on every other call use the value we get here.
         // Additionally, we cannot call this multiple times for the same directory, because the result changes
         // on subsequent runs. Therefore, we cache the value during the whole runtime of the application.
@@ -121,30 +114,6 @@ public static class SentryNative
         catch (EntryPointNotFoundException e)
         {
             Logger?.LogError(e, "Native dependency not found. Did you delete sentry.dll or move files around?");
-        }
-    }
-
-    private static string? GetInstallationId(IApplication? application = null)
-    {
-        application ??= ApplicationAdapter.Instance;
-        switch (application.Platform)
-        {
-            case RuntimePlatform.Switch:
-            case RuntimePlatform.PS5:
-            case RuntimePlatform.XboxOne:
-            case RuntimePlatform.GameCoreXboxSeries:
-            case RuntimePlatform.GameCoreXboxOne:
-                // TODO: Fetch the installation ID from sentry-native
-                // See https://github.com/getsentry/sentry-native/issues/1324
-                return null;
-
-            case RuntimePlatform.WindowsPlayer:
-            case RuntimePlatform.WindowsEditor:
-            case RuntimePlatform.LinuxPlayer:
-            case RuntimePlatform.LinuxEditor:
-                return AnalyticsSessionInfo.userId;
-            default:
-                return null;
         }
     }
 }